got rid of tests package
