+++ packages/core/echo/echo-schema/src/database.test.ts
@@ -2,13 +2,18 @@
 // Copyright 2022 DXOS.org
 //
 
-import expect from 'expect'; // TODO(burdon): Can't use chai with wait-for-expect?
+import { expect } from 'chai';
 import { inspect } from 'node:util';
 
 import { describe, test } from '@dxos/test';
+import { range } from '@dxos/util';
 
-import { create, Expando, getMeta } from './effect/reactive';
-import { createDatabase, TestBuilder } from './testing';
+import { getAutomergeObjectCore } from './automerge';
+import { create, Expando, getMeta, getSchema, getType, type ReactiveObject } from './effect/reactive';
+import { Hypergraph } from './hypergraph';
+import { clone } from './object';
+import { Filter } from './query';
+import { createDatabase, TestBuilder, Contact, Container, RecordType, Task, Todo } from './testing';
 
 // TODO(burdon): Normalize tests to use common graph data (see query.test.ts).
 
@@ -32,7 +37,6 @@ describe('Database', () => {
     await db.flush();
 
     inspect(task);
-    // console.log(task);
   });
 
   test('query by id', async () => {
@@ -47,7 +51,7 @@ describe('Database', () => {
 
     {
       const { objects } = db.query();
-      expect(objects).toHaveLength(n);
+      expect(objects.length).to.eq(n);
     }
 
     db.remove(db.query().objects[0]);
@@ -55,7 +59,7 @@ describe('Database', () => {
 
     {
       const { objects } = db.query();
-      expect(objects).toHaveLength(n - 1);
+      expect(objects.length).to.eq(n - 1);
     }
   });
 
@@ -71,22 +75,224 @@ describe('Database', () => {
     await db.flush();
 
     expectObjects(getMeta(obj).keys, [{ id: 'test-key', source: 'test' }]);
-    // TODO(mykola): Implement in automerge.
-    // expect(obj[data]).toEqual({
-    //   '@id': obj.id,
-    //   '@type': undefined,
-    //   '@model': 'dxos.org/model/document',
-    //   '@meta': {
-    //     keys: [{ id: 'test-key', source: 'test' }],
-    //   },
-    // });
+  });
+
+  test('creating objects', async () => {
+    const { db: database } = await createDbWithTypes();
+
+    const task = create(Task, { title: 'test' });
+    expect(task.title).to.eq('test');
+    expect(task.id).to.exist;
+    expect(() => getAutomergeObjectCore(task)).to.throw();
+    expect(getSchema(task)?.ast).to.eq(Task.ast);
+    expect(getType(task)?.itemId).to.eq('example.test.Task');
+
+    database.add(task);
+    await database.flush();
+    expect(getAutomergeObjectCore(task).database).to.exist;
+
+    const { objects: tasks } = database.query(Filter.schema(Task));
+    expect(tasks).to.have.length(1);
+    expect(tasks[0].id).to.eq(task.id);
+  });
+
+  test('enums', async () => {
+    const { db: database } = await createDbWithTypes();
+
+    {
+      const container = create(Container, { records: [{ type: RecordType.WORK }] });
+      await database.add(container);
+    }
+
+    {
+      const { objects } = database.query(Filter.schema(Container));
+      const [container] = objects;
+      expect(container.records).to.have.length(1);
+      expect(container.records![0].type).to.eq(RecordType.WORK);
+    }
+  });
+
+  test('dxos.schema.Expando', async () => {
+    const { db: database } = await createDbWithTypes();
+
+    {
+      const container = create(Container, { objects: [] });
+      database.add(container);
+      await database.flush();
+
+      container.objects!.push(create(Expando, { foo: 100 }));
+      container.objects!.push(create(Expando, { bar: 200 }));
+    }
+
+    {
+      const { objects } = database.query(Filter.schema(Container));
+      const [container] = objects;
+      expect(container.objects).to.have.length(2);
+      expect(container.objects![0]!.foo).to.equal(100);
+      expect(container.objects![1]!.bar).to.equal(200);
+    }
+  });
+
+  test('dxos.schema.TextObject', async () => {
+    const { db: database } = await createDbWithTypes();
+
+    {
+      const container = create(Container, { objects: [] });
+      database.add(container);
+      await database.flush();
+
+      container.objects!.push(create(Task, {}));
+      container.objects!.push(create(Contact, {}));
+    }
+
+    {
+      const { objects } = database.query(Filter.schema(Container));
+      const [container] = objects;
+      expect(container.objects).to.have.length(2);
+      expect(getType(container.objects![0])?.itemId).to.equal(Task.typename);
+      expect(getType(container.objects![1])?.itemId).to.equal(Contact.typename);
+    }
+  });
+
+  test('object fields', async () => {
+    const task = create(Task, {});
+
+    task.title = 'test';
+    expect(task.title).to.eq('test');
+    expect(getMeta(task).keys).to.have.length(0);
+
+    getMeta(task).keys.push({ source: 'example', id: 'test' });
+    expect(getMeta(task).keys).to.have.length(1);
+  });
+
+  test('clone', async () => {
+    const { db: db1 } = await createDbWithTypes();
+    const { db: db2 } = await createDbWithTypes();
+
+    const task1 = create(Task, { title: 'Main task' });
+    db1.add(task1);
+    await db1.flush();
+
+    const task2 = clone(task1);
+    expect(task2 !== task1).to.be.true;
+    expect(task2.id).to.equal(task1.id);
+    expect(task2.title).to.equal(task1.title);
+
+    db2.add(task2);
+    await db2.flush();
+    expect(task2).to.be.instanceOf(Task);
+    expect(task2.id).to.equal(task1.id);
+
+    expect(() => db1.add(task1)).to.throw;
+  });
+
+  test('operator-based filters', async () => {
+    const { db: database } = await createDbWithTypes();
+
+    database.add(create(Task, { title: 'foo 1' }));
+    database.add(create(Task, { title: 'foo 2' }));
+    database.add(create(Task, { title: 'bar 3' }));
+
+    expect(database.query(Filter.schema(Task, (task) => task.title?.startsWith('foo'))).objects).to.have.length(2);
+  });
+
+  test('typenames of nested objects', async () => {
+    const { db: database } = await createDbWithTypes();
+
+    const task = create(Task, {
+      title: 'Main task',
+      todos: [create(Todo, { name: 'Sub task' })],
+    });
+    database.add(task);
+
+    console.log(task.todos![0]);
+    expect(getType(task.todos![0] as any)?.itemId).to.eq('example.test.Task.Todo');
+    expect(JSON.parse(JSON.stringify(task.todos![0]))['@type'].itemId).to.eq('example.test.Task.Todo');
+  });
+
+  describe('object collections', () => {
+    test('assignment', async () => {
+      const root = newTask();
+      expect(root.subTasks).to.have.length(0);
+
+      range(3).forEach(() => root.subTasks!.push(newTask()));
+      root.subTasks!.push(newTask(), newTask());
+
+      expect(root.subTasks).to.have.length(5);
+      expect(root.subTasks!.length).to.eq(5);
+      expect(JSON.parse(JSON.stringify(root, undefined, 2)).subTasks).to.have.length(5);
+
+      // Iterators.
+      const ids = root.subTasks!.map((task) => task!.id);
+      root.subTasks!.forEach((task, i) => expect(task!.id).to.eq(ids[i]));
+      expect(Array.from(root.subTasks!.values())).to.have.length(5);
+
+      root.subTasks = [create(Task, {}), create(Task, {}), create(Task, {})];
+      expect(root.subTasks.length).to.eq(3);
+
+      await addToDatabase(root);
+    });
+
+    test('splice', async () => {
+      const root = newTask();
+      root.subTasks = range(3).map(newTask);
+      root.subTasks.splice(0, 2, newTask());
+      expect(root.subTasks).to.have.length(2);
+      await addToDatabase(root);
+    });
+
+    test('array of plain objects', async () => {
+      const root = create(Container, { records: [] });
+      root.records!.push({
+        title: 'test',
+        contacts: [create(Contact, { name: 'tester' })],
+      });
+      const { db } = await addToDatabase(root);
+
+      expect(root.records).to.have.length(1);
+      const queriedContainer = db.query(Filter.schema(Container)).objects[0]!;
+      expect(queriedContainer.records!.length).to.equal(1);
+      expect(queriedContainer.records![0]!.contacts![0]!.name).to.equal('tester');
+    });
+
+    test('reset array', async () => {
+      const { db, obj: root } = await addToDatabase(create(Container, { records: [] }));
+
+      root.records!.push({ title: 'one' });
+      expect(root.records).to.have.length(1);
+
+      root.records = [];
+      expect(root.records).to.have.length(0);
+      await db.flush();
+      expect(root.records).to.have.length(0);
+
+      root.records.push({ title: 'two' });
+      expect(root.records).to.have.length(1);
+      await db.flush();
+      expect(root.records).to.have.length(1);
+    });
   });
 });
 
 const expectObjects = (echoObjects: any[], expectedObjects: any) => {
-  expect(mapEchoToPlainJsObject(echoObjects)).toStrictEqual(expectedObjects);
+  expect(mapEchoToPlainJsObject(echoObjects)).to.deep.eq(expectedObjects);
 };
 
 const mapEchoToPlainJsObject = (array: any[]): any[] => {
   return array.map((o) => (Array.isArray(o) ? mapEchoToPlainJsObject(o) : { ...o }));
+};
+
+const newTask = () => create(Task, { subTasks: [] });
+
+const createDbWithTypes = async () => {
+  const graph = new Hypergraph();
+  graph.runtimeSchemaRegistry.registerSchema(Task, Contact, Container, Todo);
+  return createDatabase(graph);
+};
+
+const addToDatabase = async <T>(obj: ReactiveObject<T>) => {
+  const { db } = await createDbWithTypes();
+  db.add(obj);
+  await db.flush();
+  return { db, obj };
 };
