+++ packages/core/echo/echo-schema/src/query/query.test.ts
@@ -13,8 +13,7 @@ import { Filter } from './filter';
 import { type EchoDatabase } from '../database';
 import { Expando, create } from '../effect/reactive';
 import { type EchoReactiveObject } from '../effect/reactive';
-import { TestBuilder, createDatabase } from '../testing';
-import { Contact } from '../tests/schema';
+import { Contact, TestBuilder, createDatabase } from '../testing';
 
 const createTestObject = (idx: number, label?: string) => {
   return create(Expando, { idx, title: `Task ${idx}`, label });
