+++ packages/core/echo/echo-schema/src/testing/index.ts
@@ -3,3 +3,4 @@
 //
 
 export * from './testing';
+export * from './schema';
